requirements.txt
aiohttp
uvloop
emails
pyaml
webargs
jinja2
motor
aioredis